Move summary types to model package
