+++ model/validator_session_seq_summary.go
@@ -1,0 +1,11 @@
+package model
+
+import "github.com/figment-networks/polkadothub-indexer/types"
+
+type ValidatorSessionSeqSummary struct {
+	StashAccount string     `json:"stash_account"`
+	TimeBucket   types.Time `json:"time_bucket"`
+	UptimeAvg    float64    `json:"uptime_avg"`
+	UptimeMin    int64      `json:"uptime_min"`
+	UptimeMax    int64      `json:"uptime_max"`
+}
