+++ store/validator_session_seq.go
@@ -14,13 +14,5 @@ type ValidatorSessionSeq interface {
 	FindBySessionAndStashAccount(session int64, stash string) (*model.ValidatorSessionSeq, error)
 	FindLastByStashAccount(stashAccount string, limit int64) ([]model.ValidatorSessionSeq, error)
 	FindMostRecent() (*model.ValidatorSessionSeq, error)
-	Summarize(interval types.SummaryInterval, activityPeriods []ActivityPeriodRow) ([]ValidatorSessionSeqSummary, error)
-}
-
-type ValidatorSessionSeqSummary struct {
-	StashAccount string     `json:"stash_account"`
-	TimeBucket   types.Time `json:"time_bucket"`
-	UptimeAvg    float64    `json:"uptime_avg"`
-	UptimeMin    int64      `json:"uptime_min"`
-	UptimeMax    int64      `json:"uptime_max"`
+	Summarize(interval types.SummaryInterval, activityPeriods []ActivityPeriodRow) ([]model.ValidatorSessionSeqSummary, error)
 }
